+++ gradle-plugins/gradle.properties
@@ -8,7 +8,7 @@ kotlin.code.style=official
 dev.junit.parallel=false
 
 # Default version of Compose Libraries used by Gradle plugin
-compose.version=1.8.10+dev2320
+compose.version=1.8.10+dev2347
 # The latest version of Kotlin compatible with compose.tests.compiler.version. Used only in tests/CI.
 compose.tests.kotlin.version=2.2.0-Beta1
 # __SUPPORTED_GRADLE_VERSIONS__
